arm64: dts: exynos7904*: increase max big CPU cluster freq to 2184 MHz
Same frequency used for the big CPU cluster for the Exynos 7885.

Possible since the last commit modifying the exynos-acme driver.
Risking somewhat reiterating info in the message in the previous commit,
note that arbitrary frequency increases aren't so easily possible - With
the 7904 being a dumbing down of the 7885, it does have fully valid ECT
data up to 2184 MHz for the big CPU cluster with perfectly working
voltages and all, it's just the max frequency setting that's set to be
lower. Not that anyone should arbitrarily increase it past 2184 MHz.

Signed-off-by: K9100ii <k9100ii.ki@gmail.com>