arm64: exynos: copy a 7885 Mali GPU device tree to a 7904 one
The Mali GPU device tree config at:
  arch/arm64/boot/dts/exynos/mali/exynos7904-mali.dtsi
doesn't appear to be used for any actual device, but is somehow used by
some unused, yet still compiled, device trees. There are two others with
the same name but at different paths that are used, and they're 1:1
copies of the 7885 Mali device tree. Copy the 7885 one to that 7904 DT
just for consistency.

Signed-off-by: K9100ii <k9100ii.ki@gmail.com>